Added test suite
New test suite.

I added a new member 'rule' to the stressmodel result in order to make it easier to provide test coverage and to make sure that each test is covering the highway that is defined for that test.

I also found a small bug in parkingPresent that could affect the determination of whether there is parking spots where a bike lane is present. It currently has no effect on the output based on the current tags.