Added motor_vehicle as model tag.
One of the memory optimizations I put into place was to restrict the reference tags. This is probably no longer needed because of the way we process the ways now using two passes.  I will remove this but for now I will simply include the motor_vehicle tag to allow the model to work properly.